Skip verification of domains that are already authorized
The returned list of authorizations needed for the certificate include
authorizations that are still valid from previous runs. So rather than
retry those we can skip them instead.

This also contains a bunch of cleanups for the error handling so that
better errors are printed. Along with fixing python3 support.