Fix the DNS pre-check
I remember having a reason that it didn't give an error if the records
were completely missing, but I forgot what it was. I suspect it had to
do with split DNS configurations but thats what I added --skip for.

So unless --skip is used the DNS check now requires:
* At least one DNS server be reachable
* All reachable DNS servers must have the correct record